Добавлена фабрика вопросов, структуры вынесены в отдельные файлы
